modificaion del vec!

+++ reporte/lib.rs
@@ -30,13 +30,13 @@ use sistema_de_votacion::SistemaDeVotacionRef;
         }
 
         #[ink(message)]
-        pub fn get_reporte_de_eleccion(&self,id_eleccion:u8) -> Option<sistema_de_votacion::sistema_de_votacion::Eleccion>{
+        pub fn get_reporte_de_eleccion(&self,id_eleccion:i16) -> Option<sistema_de_votacion::sistema_de_votacion::Eleccion>{
             let eleccion = self.sistema_de_votacion.get_reporte_de_eleccion(id_eleccion);
             eleccion
         }
 
         #[ink(message)]
-        pub fn reporte_de_eleccion(&self,id_eleccion:u8) -> Votantes{
+        pub fn reporte_de_eleccion(&self,id_eleccion:i16) -> Votantes{
             let eleccion = self.sistema_de_votacion.get_reporte_de_eleccion(id_eleccion);
             let votantes=Votantes::new();
             match eleccion {
